[feat] #33 발매 정보 뷰 아이템 클릭 시 이동 및 스크랩
